better error messages, better `parent_album` logic
error messages are now super explicit about what failed. the encoding functions now always check `self.parent_album: Option<(usize, usize)>` (which are just indeces) for whether or not a parent album exists, instead of being a parameter to the function. so the branch we take is always tied to the attribute, not the parameter. :)